Improve CI investigation
* Improve task naming consistency
* Fix issue with KUBECONFIG not being output in case of error
* Output KUBECONFIG early in the process to allow real time investigations
* Fix an issue that would trigger the tests to fail if the user deleted
  the destroy-cluster.txt before the tests finished

+++ ci/images/ci-runner/hack/sidecar/bin/plnsvc_test.sh
@@ -8,5 +8,9 @@ echo "Start executing pipeline cases ..."
 TEST_DIR=$(find "$PWD" -type f -name test.sh -exec dirname {} +)
 "$TEST_DIR/test.sh" --kubeconfig "$KUBECONFIG"
 
-# If the tests are successful, the cluster can be destroyed right away
-rm "$PWD/destroy-cluster.txt"
+
+# In case the user deleted the file early when they expected a failure
+if [ -e "$PWD/destroy-cluster.txt" ]; then
+    # If the tests are successful, the cluster can be destroyed right away
+    rm "$PWD/destroy-cluster.txt"
+fi
